move the loki upload to an invariant serializer
